+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 
 import './globals.css';
-import Header from './(layout)/Header';
+import Header from '../components/layouts/Header';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
